Break down BenchmarkData file into separate files
