+++ IronBenchmarks.Reporting.Tests/ReportGeneratorTests.cs
@@ -1,4 +1,5 @@
 ﻿using IronBenchmarks.Reporting.Configuration;
+using IronBenchmarks.Reporting.ReportData;
 using IronXL;
 
 namespace IronBenchmarks.Reporting.Tests
