+++ IronBenchmarks.ReportsEngine/ReportData/BenchmarkData.cs
@@ -4,7 +4,7 @@ using System.ComponentModel;
 using System.Linq;
 using static BenchmarkDotNet.Reports.SummaryTable;
 
-namespace IronBenchmarks.Reporting
+namespace IronBenchmarks.Reporting.ReportData
 {
     public class BenchmarkData
     {
@@ -92,44 +92,5 @@ namespace IronBenchmarks.Reporting
         {
             return DataEntries.FirstOrDefault().GetKeys().Length;
         }
-    }
-
-    public class BenchmarkDataEntry
-    {
-        private readonly Dictionary<string, double> data = new Dictionary<string, double>();
-
-        public string Name { get; }
-        public int Count => data.Count;
-
-        public BenchmarkDataEntry(string name)
-        {
-            Name = name;
-        }
-
-        public void Add(string title, double value)
-        {
-            data.Add(title, value);
-        }
-
-        public string[] GetKeys()
-        {
-            return data.Keys.ToArray();
-        }
-
-        public double this[string key]
-        {
-            get
-            {
-                return data[key];
-            }
-        }
-    }
-
-    public enum ReportDataType
-    {
-        [Description("Performance")]
-        MeanTime,
-        [Description("Memory Allocation")]
-        MemoryAlloc
     }
 }
