+++ IronBenchmarks.ReportsEngine/ReportData/ReportDataType.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+
+namespace IronBenchmarks.Reporting.ReportData
+{
+    public enum ReportDataType
+    {
+        [Description("Performance")]
+        MeanTime,
+        [Description("Memory Allocation")]
+        MemoryAlloc
+    }
+}
